fix(edge): move chisel crypto into a separate directory EE-5099
