+++ api/chisel/crypto/crypto.go
@@ -1,4 +1,4 @@
-package chisel
+package crypto
 
 import (
 	"crypto/ecdsa"
@@ -23,7 +23,7 @@ func GenerateGo119CompatibleKey(seed string) ([]byte, error) {
 	}
 	b, err := x509.MarshalECPrivateKey(priv)
 	if err != nil {
-		return nil, fmt.Errorf("Unable to marshal ECDSA private key: %v", err)
+		return nil, fmt.Errorf("Unable to marshal ECDSA private key: %w", err)
 	}
 	return pem.EncodeToMemory(&pem.Block{Type: "EC PRIVATE KEY", Bytes: b}), nil
 }
